fix: clickable sort instead of drag and drop
