+++ src/screens/Profile/FavoriteList/SortFavorites.tsx
@@ -1,8 +1,5 @@
 import {Confirm} from '@atb/assets/svg/icons/actions';
-import SvgDragHandle from '@atb/assets/svg/icons/actions/DragHandle';
 import Button from '@atb/components/button';
-import {FavoriteItem, Section} from '@atb/components/sections';
-import ThemeIcon from '@atb/components/theme-icon';
 import {useFavorites} from '@atb/favorites';
 import MessageBox from '@atb/components/message-box';
 import {TabNavigatorParams} from '@atb/navigation/TabNavigator';
@@ -11,16 +8,12 @@ import {FavoriteListTexts, useTranslation} from '@atb/translations';
 import {CompositeNavigationProp} from '@react-navigation/native';
 import {StackNavigationProp} from '@react-navigation/stack';
 import React, {useState} from 'react';
-import {PanGestureHandler} from 'react-native-gesture-handler';
-import Animated from 'react-native-reanimated';
 import {ProfileStackParams} from '..';
-import {SortableList} from './SortableList';
-import SortableListFallback from './SortableListFallback';
+import SortableList from './SortableList';
 import useFontScale from '@atb/utils/use-font-scale';
 import FullScreenHeader from '@atb/components/screen-header/full-header';
 import {View} from 'react-native';
 import FullScreenFooter from '@atb/components/screen-footer/full-footer';
-import {useAccessibilityContext} from '@atb/AccessibilityContext';
 
 export type ProfileScreenNavigationProp = StackNavigationProp<
   ProfileStackParams,
@@ -42,9 +35,6 @@ export default function SortableFavoriteList({navigation}: ProfileScreenProps) {
   const items = favorites ?? [];
   const [sortedItems, setSortedItems] = useState(items);
   const [error, setError] = useState<string | null>(null);
-  const a11yContext = useAccessibilityContext();
-  const fontScale = useFontScale();
-  const minHeight = 40 + 12 * fontScale;
   const {t} = useTranslation();
 
   const saveOrder = async () => {
@@ -67,40 +57,7 @@ export default function SortableFavoriteList({navigation}: ProfileScreenProps) {
         <MessageBox type="error" message={error} containerStyle={style.error} />
       )}
 
-      {a11yContext.isScreenReaderEnabled ? (
-        <SortableListFallback data={sortedItems} onSort={setSortedItems} />
-      ) : (
-        <SortableList
-          data={items}
-          // @TODO Find solution for not hardcoding this.
-          rowHeight={minHeight}
-          onSort={setSortedItems}
-          containerStyle={style.orderContainer}
-          renderRow={(data, index, state, onGestureEvent) => {
-            let opacity = state === 'placeholder' ? 0 : 1;
-            if (state === 'dragging') {
-              opacity -= 0.5;
-            }
-            return (
-              <PanGestureHandler
-                maxPointers={1}
-                onGestureEvent={onGestureEvent}
-                onHandlerStateChange={onGestureEvent}
-              >
-                <Animated.View style={{flex: 1, opacity}}>
-                  <Section withPadding>
-                    <FavoriteItem
-                      favorite={data}
-                      icon={<ThemeIcon svg={SvgDragHandle} />}
-                    />
-                  </Section>
-                </Animated.View>
-              </PanGestureHandler>
-            );
-          }}
-          indexToKey={(i) => items[i].id}
-        />
-      )}
+      <SortableList data={sortedItems} onSort={setSortedItems} />
 
       <FullScreenFooter>
         <Button
